Revert "drop lines for python 3.5"
This reverts commit a60b706790a270d5429ec9c6d857a8be8553e5dc.

+++ composeml/label_times/object.py
@@ -67,7 +67,8 @@ class LabelTimes(DataFrame):
             value = self.groupby('cutoff_time')
             value = value[self.label_name].count()
 
-        value = value.cumsum()
+        value = value.cumsum()  # In Python 3.5, these values automatically convert to float.
+        value = value.astype('int')
         return value
 
     def describe(self):
